fix chat message formatting
Updated project version to 4.7.7 in build.gradle.kts and plugin.yml. Fixed chat message formatting in PlayerChatEvent to apply color codes to the entire prefix and player name, and corrected message serialization to use the formatted message. Minor code cleanup in HuskHomesHook and FloodgateHook, and switched to PaperCommandManager in UltimateTeams.

#101